Проверка данных от пользователя; Array_column; Порядок в коде: Разделение на файлы, добавление функций, избавление от дблирования
