+++ selecttasks.php
@@ -1,0 +1,14 @@
+<?php
+    //Функция вывода всех тасков
+    function show_tasks ($user_id) {
+        $sql = 'SELECT tasks.id, status, task_name, run_to, project_id FROM tasks ' . 'JOIN projects ON tasks.project_id = projects.id ' . 'WHERE tasks.user_id = "' . $user_id . '"';
+
+        return $sql;
+    }
+
+    function show_cat_tasks ($user_id, $cat_id) {
+        $sql = 'SELECT tasks.id, status, task_name, run_to, project_id FROM tasks ' . 'JOIN projects ON tasks.project_id = projects.id ' . 'WHERE tasks.user_id = "' . $user_id . '" AND project_id = "' . $cat_id . '"';
+
+        return $sql;
+    }
+?>
